Version of PELICAN that sort of runs through Vitis AI, along with instructions
The hardware/ directory contains notes for setting this up. You need the
pre-trained model files, which are not included.

This is kind of a... mess... right now, but this is where I got to last week
in preparation for the talk. Hopefully it's a useful set of notes for someone
to build off of.